Improve anti-second run

+++ src/client/src/main.rs
@@ -125,7 +125,7 @@ fn should_run() -> bool {
     system.refresh_all();
 
     for (pid, proc) in system.get_process_list() {
-        if pid == std::process::id() {
+        if pid as i32 == std::process::id() as i32 {
             continue;
         }
